Resolved Merge Conflicts from latest pull
